fix: remove double cast in core calculations DDR4, DDR5

+++ src/DRAMPower/DRAMPower/standards/ddr4/core_calculation_DDR4.cpp
@@ -27,7 +27,7 @@ namespace DRAMPower {
 	}
 
 	inline double Calculation_DDR4::E_RD(double VDD, double IDD4_R, double I_B, double t_CK, std::size_t BL, std::size_t DR, uint64_t N_RD) const {
-        return VDD * (IDD4_R - I_B) * (double(BL) / DR) * t_CK * N_RD;
+        return VDD * (IDD4_R - I_B) * (BL / DR) * t_CK * N_RD;
 	}
 
 	inline double Calculation_DDR4::E_WR(double VDD, double IDD4_W, double I_B, double t_CK, std::size_t BL, std::size_t DR, uint64_t N_WR) const {
